+++ projects/framework-photoshop/source/ftrack_framework_photoshop/__init__.py
@@ -34,6 +34,8 @@ from ftrack_framework_core.client import Client
 from ftrack_framework_core.configure_logging import configure_logging
 from ftrack_framework_core import registry
 
+from ftrack_utils.session import create_api_session
+
 
 # Evaluate version and log package version
 try:
@@ -145,16 +147,26 @@ def bootstrap_integration(framework_extensions_path):
         f' {framework_extensions_path}'
     )
 
-    session = ftrack_api.Session(auto_connect_event_hub=True)
+    session = create_api_session(auto_connect_event_hub=True)
 
-    event_manager = EventManager(session=session)
+    event_manager = EventManager(
+        session=session, mode=constants.event.LOCAL_EVENT_MODE
+    )
 
     registry_instance = registry.Registry()
     registry_instance.scan_extensions(paths=framework_extensions_path)
 
-    Host(event_manager, registry=registry_instance)
+    Host(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=invoke_in_qt_main_thread,
+    )
 
-    client_instance = Client(event_manager, registry=registry_instance)
+    client_instance = Client(
+        event_manager,
+        registry=registry_instance,
+        run_in_main_thread_wrapper=invoke_in_qt_main_thread,
+    )
 
     # Init tools
     dcc_config = registry_instance.get_one(